Audio/ValKit: Added support for CRC32 checksum calculation; now receiving data also works with bigger chunks [build fix]. bugref:10008
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@89807 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Audio/AudioTestServiceClient.cpp
@@ -162,7 +162,7 @@ static int audioTestSvcClientRecvReply(PATSCLIENT pClient, PATSSRVREPLY pReply, 
                             rc = VERR_NET_PROTOCOL_ERROR;
                     }
 
-                    cbPayloadRead += cbRead;
+                    cbPayloadRead += (uint32_t)cbRead;
                     Assert(cbPayloadRead <= pReply->cbPayload);
                 }
             }
